pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cac-content-mcp-server"
version = "0.1.0"
description = "MCP server for ComplianceAsCode/content project enabling AI-powered security compliance automation"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "ComplianceAsCode Contributors"}
]
keywords = ["mcp", "security", "compliance", "scap", "ansible"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    "mcp>=0.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "aiofiles>=23.0.0",
    "GitPython>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML",
]
http = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
]

[project.scripts]
cac-mcp-server = "cac_mcp_server.__main__:main"

[project.urls]
Homepage = "https://github.com/ComplianceAsCode/cac-content-mcp-server"
Documentation = "https://github.com/ComplianceAsCode/cac-content-mcp-server/blob/main/README.md"
Repository = "https://github.com/ComplianceAsCode/cac-content-mcp-server"
Issues = "https://github.com/ComplianceAsCode/cac-content-mcp-server/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
cac_mcp_server = ["config/defaults.yaml", "prompts/**/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=cac_mcp_server",
    "--cov-report=term-missing",
]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by black)
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
